[157] way too many things done.
You'll see al the changes inside the changelog if this snapshot becomes a thing, I mean... there's no one here to see, I can wait...
