+++ java/src/moonaFramework/annotations/Deadlined.java
@@ -1,4 +1,4 @@
-package moonaFramework;
+package moonaFramework.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.RetentionPolicy;
