+++ java/src/moonaFramework/annotations/Timeless.java
@@ -1,7 +1,6 @@
-package moonaFramework;
+package moonaFramework.annotations;
 
-import static java.lang.annotation.ElementType.CONSTRUCTOR;
-import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.ElementType.*;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 import java.lang.annotation.Retention;
