+++ java/src/moonaFramework/process/ProcessHandler.java
@@ -1,10 +1,9 @@
 package moonaFramework.process;
 
-import java.util.stream.Stream;
-import moonaFramework.Timeless;
-
 public interface ProcessHandler {
 
+	void MainStart(Process p);
+	
 	void Provide(Process p);
 	
 	void Await(Process p);
@@ -18,12 +17,4 @@ public interface ProcessHandler {
 	
 	void Terminate(Process p);
 	void Interrupt(Process p);
-	
-	static void initializer(Process p) {
-		if (Stream.of(p.getClass().getMethods())
-				.filter(method -> method.getName() == "initialize")
-				.findAny().get().getAnnotation(Timeless.class) != null) {
-			p.getClock().stasys();
-		}
-	}
 }
