+++ java/src/moonaFramework/reflection/Mirror.java
@@ -1,0 +1,5 @@
+package moonaFramework.reflection;
+
+public class Mirror {
+
+}
