+++ java/src/test/Test.java
@@ -1,27 +1,39 @@
 package test;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import moonaFramework.Benchmark;
-import moonaFramework.Moona;
+
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+import java.util.Comparator;
+import java.util.stream.Stream;
+
+import moonaFramework.*;
 import moonaFramework.ProcessCondition;
 import moonaFramework.Status;
+import moonaFramework.annotations.*;
 import moonaFramework.process.*;
 import moonaFramework.util.Condition;
 
 @SuppressWarnings("unused")
 public class Test {
 	
-	static Moona m;
-	
-	static Daemon d = new Daemon(m) {
+	static Phase p = new Phase();
+
+	static Task t = new Task() {
+		@Timeless
+		public void initialize() {
+			while (true) {
+				
+			}
+		}
 		public void update() {
-			
+			System.out.println("AAAAAAAAA");
 		}
 	};
 	
 	public static void main(String[] args) throws InterruptedException {
+		Moona.Init();
 		
+		p.Start(t);
 	}
 }
